chore: resource and abi change

+++ contracts/Migrations.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.4.26 <0.9.0;
 
 contract Migrations {
   address public owner = msg.sender;
-  uint public last_completed_migration;
+  uint256 public last_completed_migration;
 
   constructor() public {
     owner = msg.sender;
@@ -18,7 +18,7 @@ contract Migrations {
     _;
   }
 
-  function setCompleted(uint completed) public restricted {
+  function setCompleted(uint256 completed) public restricted {
     last_completed_migration = completed;
   }
